+++ server/main.ts
@@ -1,5 +1,6 @@
 import { migrateSettings } from './lib/migration/settings'
 import { initSettings } from './lib/settings'
+import { initCustomFields } from './lib/custom-fields'
 import { initRouters } from './lib/routers/index'
 import { ensureProsodyRunning, ensureProsodyNotRunning } from './lib/prosody/ctl'
 import decache from 'decache'
@@ -19,6 +20,7 @@ async function register (options: RegisterServerOptions): Promise<any> {
   await migrateSettings(options)
 
   await initSettings(options)
+  await initCustomFields(options)
   await initRouters(options)
 
   await ensureProsodyRunning(options)